refactor(utils): split web_scraper into package
